Make table navigation compatible with CodeMirror 6
