+++ package.json
@@ -30,7 +30,9 @@
     "ts-loader": "^9.3.1",
     "typescript": "^4.8.2",
     "webpack": "^5.74.0",
-    "webpack-cli": "^4.10.0"
+    "webpack-cli": "^4.10.0",
+    "@codemirror/view": "6.24.1",
+    "@codemirror/state": "6.4.1"
   },
   "dependencies": {
     "@felisdiligens/md-table-tools": "^0.1.0",
